// src/types/anime.schedule.ts
export type IScheduleArray = ISchedule[];

export interface ISchedule {
  day: number;
  list: Title[];
}

export interface Title {
  id: number;
  code: string;
  names: Names;
  franchises: Franchise[];
  announce?: string;
  status: Status;
  posters: Posters;
  updated: number;
  last_change: number;
  type: Type;
  genres: string[];
  team: Team;
  season: Season;
  description?: string;
  in_favorites: number;
  blocked: Blocked;
  player: Player;
  torrents: Torrents;
}